Refactoring top page
