+++ animations/page_transitions.styl
@@ -6,5 +6,5 @@
     background-color black
     width 100%
     height 0
-    opacity 0.5
+    opacity 0.3
     bottom 0
