+++ assets/css/main.styl
@@ -2,7 +2,8 @@ body
   background-color black
   color white
   font-family "DIN Next W04","Helvetica Neue",Helvetica,Arial,sans-serif
-  margin: 0px;
+  margin 0px;
+  text-align center
 
 a:visited,
 a:link
