+++ components/common/WaitingScroll.vue
@@ -5,34 +5,36 @@
 </template>
 
 <style lang="stylus" scoped>
+.preloader-nav
+  text-align center
+
 .preloader-nav::after
-  content ''
-  display block
-  position relative
-  width 8px
-  height 8px
+  animation touch-dot 1.8s cubic-bezier(0.23, 1, 0.32, 1) infinite
   background-color white
   border-radius 100%
-  margin 0 auto
   bottom 0
+  content ''
+  display block
+  height 8px
   line-height 0
+  margin 0 auto
+  position relative
   text-align center
-  animation touch-dot 1.8s cubic-bezier(0.23, 1, 0.32, 1) infinite
   transition opacity 0.8s linear 2.5s,transform 1.5s cubic-bezier(0.23, 1, 0.32, 1) 2.5s
+  width 8px
 
 .preloader-nav
-  font-style italic
-  font-family "ff-tisa-web-pro",serif
-  font-weight 50
-  line-height: 26px
-  position absolute
   -webkit-font-smoothing antialiased
-  font-size 12px
-  opacity 0
   bottom 40px
+  font-family "ff-tisa-web-pro",serif
+  font-size 12px
+  font-style italic
+  font-weight 50
   left 0
-  right 0
+  line-height: 26px
   margin 0 auto
+  position absolute
+  right 0
   width 200px
 
 @keyframes { touch-dot }
