+++ components/pages/top/TitleText.vue
@@ -1,0 +1,17 @@
+<template lang="pug">
+  .root
+    span
+      | katsuma narisawa
+</template>
+
+<style lang="stylus" scoped>
+span
+  display block
+  text-transform uppercase
+  font-size 20px
+  letter-spacing 4px
+  text-align center
+  position relative
+  font-weight 150
+  padding-bottom 15px
+</style>
